Rework test files to be able to test proto2
