+++ test/python/conftest.py
@@ -1,14 +1,7 @@
-from flask import request
-from pilgrim3.app import app as pilgrim_app
 from pytest import fixture
 from selenium import webdriver
-from socket import error as socket_error
-from threading import Thread
-
 import httplib
 import logging
-import os
-import time
 
 FAIL = '\033[91m'
 ENDC = '\033[0m'
@@ -50,42 +43,6 @@ def ghostdriver_log_path():
     yield "log/ghostdriver.log"
 
 
-@fixture(scope="session")
-def app():
-    # OMG there is no signaling of threads in python 2.x - so
-    # the only way to tell the flask app to shutdown is this.
-    # I searched far and wide, jesus christ. http://flask.pocoo.org/snippets/67/
-    #
-    # The app thread should be started as a damon so we get a kill -9 after the timeout expires
-    @pilgrim_app.route('/shutdown', methods=['GET'])
-    def shutdown():
-        request.environ.get('werkzeug.server.shutdown')()
-        return "shutdown"
-
-    @pilgrim_app.route('/booted', methods=['GET'])
-    def booted():
-        return "booted"
-
-    pilgrim_app.config['proto-bundle'] = os.path.abspath("test/support/build/types.build")
-    yield pilgrim_app
-
-
-@fixture(scope="session")
-def server_thread(app, hostname, port, client_provider):
-    server = Thread(target=lambda: app.run(host=hostname, port=port, threaded=True))
-    server.daemon = True
-    yield server
-    try:
-        client_provider().request("GET", "/shutdown")
-    except (httplib.CannotSendRequest, socket_error):
-        print FAIL + "Error shutting down" + ENDC
-
-    server.join(1)
-    if server.is_alive():
-        print FAIL + "Unable to gracefully shutdown server"
-        print "Resources may not have been released" + ENDC
-
-
 # use a provider to make connection to server after it has been booted
 @fixture(scope="session")
 def client_provider(hostname, port, timeout):
@@ -93,27 +50,6 @@ def client_provider(hostname, port, timeout):
         return httplib.HTTPConnection(hostname, port, timeout=timeout)
 
     yield get
-
-
-@fixture(scope="session")
-def server_did_boot(server_thread, hostname, port, boot_timeout):
-    server_thread.start()
-    retry_count = 0
-    booted = False
-    while True:
-        try:
-            boot_client = httplib.HTTPConnection(hostname, port, timeout=boot_timeout)
-            boot_client.request("GET", "/booted")
-            boot_client.getresponse()
-            booted = True
-        except (httplib.CannotSendRequest, socket_error):
-            print "error trying to startup:, connecting again"
-            time.sleep(0.1)
-            retry_count = retry_count + 1
-            if retry_count < 200:  # 20 seconds
-                continue
-        break
-    yield booted
 
 
 # Reset app logs every time so each test gets fresh logs when failure occurs
@@ -142,6 +78,7 @@ def driver(ghostdriver_log_path):
 def debug_logs(app_log_path, ghostdriver_log_path):
     yield (("App Log", app_log_path), ("PhantomJS Log", ghostdriver_log_path))
 
+
 @fixture
 def print_debug_func(debug_logs):
     def print_logs():
@@ -150,4 +87,5 @@ def print_debug_func(debug_logs):
             with open(debug_log[1], "r") as f:
                 print(f.read())
             print("******************")
+
     yield print_logs
